some safety checks added for model and lidar-camera fusion

+++ include/cone_detection/cone_detection.hpp
@@ -69,7 +69,7 @@ struct ConeInfo {
 typedef struct {
     float max_len;
     float min_len;
-    float interp_factor;
+    int interp_factor;
 } FusionParams;
 
 /**
